+++ src/CrowdControlModSystem.cs
@@ -5,6 +5,7 @@ using Terraria;
 using Terraria.Graphics;
 using Terraria.ID;
 using Terraria.ModLoader;
+using Terraria.ModLoader.IO;
 
 namespace CrowdControlMod;
 
@@ -34,6 +35,12 @@ public sealed class CrowdControlModSystem : ModSystem
 
     /// <inheritdoc cref="PostUpdateNPCs" />
     public static event Action? PostUpdateNPCsHook;
+
+    /// <inheritdoc cref="SaveWorldData" />
+    public static event Action<TagCompound>? SaveWorldDataHook;
+
+    /// <inheritdoc cref="LoadWorldData" />
+    public static event Action<TagCompound>? LoadWorldDataHook; 
 
     #endregion
 
@@ -71,6 +78,16 @@ public sealed class CrowdControlModSystem : ModSystem
     public override void PostUpdateNPCs()
     {
         PostUpdateNPCsHook?.Invoke();
+    }
+
+    public override void SaveWorldData(TagCompound tag)
+    {
+        SaveWorldDataHook?.Invoke(tag);
+    }
+
+    public override void LoadWorldData(TagCompound tag)
+    {
+        LoadWorldDataHook?.Invoke(tag);
     }
 
     #endregion